+++ packages/ink-editor/src/utils/configureLanguage.ts
@@ -1,5 +1,6 @@
 import { monaco } from 'react-monaco-editor';
-import { WorldState } from '../../../../pkg/rust_analyzer_wasm';
+import { Thenable } from 'monaco-editor/esm/vs/editor/editor.api';
+import { WorldState } from '../../pkg/rust_analyzer_wasm';
 
 const modeId = 'ra-rust';
 
@@ -42,8 +43,30 @@ export const configureLanguage = (worldState: WorldState, allTokens: Token[]) =>
   });
 
   monaco.languages.registerHoverProvider(modeId, {
-    provideHover: (_, pos) => worldState.hover(pos.lineNumber, pos.column),
+    provideHover: (_, pos) => {
+      const res = worldState.hover(pos.lineNumber, pos.column) as Thenable<monaco.languages.Hover>;
+      const mod = res.then(content => {
+        if (content.range) {
+          return {
+            contents: [
+              {
+                ...content.contents[0],
+                value: JSON.parse(
+                  JSON.stringify(content.contents[0]?.value)
+                    .replaceAll(/```(.*?)```/g, '```rust$1```')
+                    .replaceAll('rustrust', 'rust')
+                ),
+                supportThemeIcons: true,
+              },
+            ],
+            range: content.range,
+          } as monaco.languages.Hover;
+        }
+      });
+      return mod;
+    },
   });
+
   monaco.languages.registerCodeLensProvider(modeId, {
     async provideCodeLenses(m) {
       const code_lenses: Array<CodeLensSymbol> = await worldState.code_lenses();
