+++ packages/ink-editor/src/utils/startRustAnalyzer.ts
