+++ packages/playground/src/app/index.tsx
@@ -1,16 +1,33 @@
 import { Console } from './Console';
-import { Editor } from './Editor';
+import { InkEditor, exampleCode } from '@paritytech/ink-editor';
 import { Layout } from './Layout';
 import { Header } from './Header';
-import { AppProvider } from '~/context/app/';
+import { AppContext, AppProvider } from '~/context/app/';
 import { MessageProvider } from '../context/messages/';
-import { ReactElement } from 'react';
+import { ReactElement, useContext, useState } from 'react';
+import { Dispatch, State } from '~/context/app/reducer';
 
 const App = (): ReactElement => {
+  const [code, setCode] = useState(exampleCode);
+  const [state, dispatch]: [State, Dispatch] = useContext(AppContext);
+
   return (
     <AppProvider>
       <MessageProvider>
-        <Layout Header={Header} Editor={Editor} Console={Console} />
+        <Layout
+          Header={Header}
+          Editor={() => (
+            <InkEditor
+              code={code}
+              onCodeChange={setCode}
+              numbering={state.numbering}
+              darkmode={state.darkmode}
+              minimap={state.minimap}
+              setURI={uri => dispatch({ type: 'SET_URI', payload: uri })}
+            />
+          )}
+          Console={Console}
+        />
       </MessageProvider>
     </AppProvider>
   );
